refactor: split service ownership boundaries (#1729)
