+++ tests/test_service.py
@@ -3,7 +3,6 @@ import hashlib
 import hmac
 import json
 import os
-import tempfile
 import unittest
 from collections.abc import Mapping
 from pathlib import Path
@@ -17,7 +16,6 @@ from click.testing import CliRunner
 from control_plane.cli import main
 from control_plane import live_target_runtime as control_plane_live_target_runtime
 from control_plane import secrets as control_plane_secrets
-from control_plane import service as control_plane_service
 from control_plane.notifications import public_discord_url_error, public_url_error
 from control_plane.contracts.environment_inventory import EnvironmentInventory
 from control_plane.contracts.every_code_notifications import (
@@ -78,9 +76,9 @@ from control_plane.contracts.runner_lane_registration import (
     RunnerLaneRegistrationRequest,
     plan_runner_lane_registration,
 )
-from control_plane.service import (
+from control_plane.every_code_github_webhook import handle_every_code_github_webhook_request
+from control_plane.drivers.generic_web_preview_dispatch import (
     GenericWebPreviewVerificationRequest,
-    handle_every_code_github_webhook_request,
 )
 from control_plane.http_app import LaunchplaneAuthzPolicyRuntime
 from control_plane.http_app import create_launchplane_fastapi_app
@@ -89,7 +87,7 @@ from control_plane.generic_web_promotion_http import (
     GenericWebProdPromotionResponse,
     GenericWebPromotionWorkflowResponse,
 )
-from control_plane.drivers import generic_web_preview_dispatch, native_routes
+from control_plane.drivers import generic_web_preview_dispatch
 from control_plane.service_auth import (
     BearerIdentityConfig,
     GitHubActionsIdentity,
@@ -2083,1075 +2081,6 @@ class LaunchplaneServiceTests(unittest.TestCase):
         self.assertEqual(request_schema["title"], "MergeTrainPolicyImportEnvelope")
         self.assertEqual(request_schema["additionalProperties"], False)
 
-    def test_every_code_github_webhook_creates_work_request(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        policy = LaunchplaneAuthzPolicy.model_validate(
-            {
-                "github_actions": [
-                    {
-                        "repository": "cbusillo/launchplane",
-                        "workflow_refs": [
-                            "cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main"
-                        ],
-                        "event_names": ["workflow_dispatch"],
-                        "products": ["launchplane"],
-                        "contexts": ["launchplane"],
-                        "actions": ["every_code_work_request.read"],
-                    }
-                ]
-            }
-        )
-        identity = _identity(
-            repository="cbusillo/launchplane",
-            workflow_ref="cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main",
-            event_name="workflow_dispatch",
-        )
-        webhook_payload = _every_code_github_issue_labeled_payload(label="EVERY-CODE")
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(identity),
-                authz_policy=policy,
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            webhook_status, webhook_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=webhook_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-1",
-                    "X-Hub-Signature-256": _github_webhook_signature(webhook_payload, secret),
-                },
-            )
-            work_requests = FilesystemRecordStore(state_dir).list_every_code_work_request_records(
-                state="queued"
-            )
-
-        self.assertEqual(webhook_status, 202)
-        self.assertFalse(webhook_response["deduped"])
-        self.assertEqual(webhook_response["records"]["state"], "queued")
-        self.assertEqual(webhook_response["github_delivery_id"], "delivery-1")
-        self.assertEqual(len(work_requests), 1)
-        request = work_requests[0]
-        self.assertEqual(request.source, "github_issue_label")
-        self.assertEqual(request.repository, "cbusillo/code")
-        self.assertEqual(request.issue_number, 123)
-        self.assertEqual(request.trigger_label, "every-code")
-        self.assertEqual(request.trigger_actor, "cbusillo")
-        self.assertEqual(request.github_delivery_id, "delivery-1")
-
-    def test_every_code_github_webhook_dedupes_existing_request(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        policy = LaunchplaneAuthzPolicy.model_validate(
-            {
-                "github_actions": [
-                    {
-                        "repository": "cbusillo/launchplane",
-                        "workflow_refs": [
-                            "cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main"
-                        ],
-                        "event_names": ["workflow_dispatch"],
-                        "products": ["launchplane"],
-                        "contexts": ["launchplane"],
-                        "actions": [
-                            "every_code_work_request.read",
-                            "every_code_work_request.claim",
-                        ],
-                    }
-                ]
-            }
-        )
-        identity = _identity(
-            repository="cbusillo/launchplane",
-            workflow_ref="cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main",
-            event_name="workflow_dispatch",
-        )
-        webhook_payload = _every_code_github_issue_labeled_payload()
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(identity),
-                authz_policy=policy,
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            first_status, first_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=webhook_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-1",
-                    "X-Hub-Signature-256": _github_webhook_signature(webhook_payload, secret),
-                },
-            )
-            request_id = str(first_payload["records"]["request_id"])
-            claim_status, _claim_payload = _claim_every_code_work_request_in_filesystem(
-                state_dir,
-                request_id,
-            )
-            second_status, second_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=webhook_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-2",
-                    "X-Hub-Signature-256": _github_webhook_signature(webhook_payload, secret),
-                },
-            )
-            stored_request = FilesystemRecordStore(state_dir).read_every_code_work_request_record(
-                request_id
-            )
-
-        self.assertEqual(first_status, 202)
-        self.assertEqual(claim_status, 202)
-        self.assertEqual(second_status, 202)
-        self.assertTrue(second_payload["deduped"])
-        self.assertEqual(second_payload["records"]["state"], "claimed")
-        self.assertEqual(stored_request.state, "claimed")
-        self.assertEqual(stored_request.github_delivery_id, "delivery-1")
-
-    def test_every_code_github_webhook_dedupes_finished_request(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        policy = LaunchplaneAuthzPolicy.model_validate(
-            {
-                "github_actions": [
-                    {
-                        "repository": "cbusillo/launchplane",
-                        "workflow_refs": [
-                            "cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main"
-                        ],
-                        "event_names": ["workflow_dispatch"],
-                        "products": ["launchplane"],
-                        "contexts": ["launchplane"],
-                        "actions": [
-                            "every_code_work_request.read",
-                            "every_code_work_request.claim",
-                            "every_code_work_request.update",
-                        ],
-                    }
-                ]
-            }
-        )
-        identity = _identity(
-            repository="cbusillo/launchplane",
-            workflow_ref="cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main",
-            event_name="workflow_dispatch",
-        )
-        webhook_payload = _every_code_github_issue_labeled_payload()
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(identity),
-                authz_policy=policy,
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            first_status, first_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=webhook_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-1",
-                    "X-Hub-Signature-256": _github_webhook_signature(webhook_payload, secret),
-                },
-            )
-            request_id = str(first_payload["records"]["request_id"])
-            claim_status, _claim_payload = _claim_every_code_work_request_in_filesystem(
-                state_dir,
-                request_id,
-            )
-            done_status, _done_payload = _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                request_id,
-                state="done",
-                result_pr_url="https://github.com/cbusillo/code/pull/26",
-            )
-            second_status, second_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=webhook_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-2",
-                    "X-Hub-Signature-256": _github_webhook_signature(webhook_payload, secret),
-                },
-            )
-
-        self.assertEqual(first_status, 202)
-        self.assertEqual(claim_status, 202)
-        self.assertEqual(done_status, 202)
-        self.assertEqual(second_status, 202)
-        self.assertTrue(second_payload["deduped"])
-        self.assertEqual(second_payload["records"]["state"], "done")
-        self.assertEqual(
-            second_payload["result"]["request"]["result_pr_url"],
-            "https://github.com/cbusillo/code/pull/26",
-        )
-
-    def test_every_code_issue_close_marks_linked_request_done(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        policy = LaunchplaneAuthzPolicy.model_validate(
-            {
-                "github_actions": [
-                    {
-                        "repository": "cbusillo/launchplane",
-                        "workflow_refs": [
-                            "cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main"
-                        ],
-                        "event_names": ["workflow_dispatch"],
-                        "products": ["launchplane"],
-                        "contexts": ["launchplane"],
-                        "actions": [
-                            "every_code_work_request.read",
-                            "every_code_work_request.claim",
-                            "every_code_work_request.update",
-                        ],
-                    }
-                ]
-            }
-        )
-        identity = _identity(
-            repository="cbusillo/launchplane",
-            workflow_ref="cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main",
-            event_name="workflow_dispatch",
-        )
-        issue_payload = _every_code_github_issue_labeled_payload()
-        close_payload = _every_code_github_issue_labeled_payload(
-            action="closed",
-            closed_at="2026-05-06T16:20:00Z",
-            state_reason="completed",
-        )
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(identity),
-                authz_policy=policy,
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            create_status, create_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = str(create_payload["records"]["request_id"])
-            _claim_every_code_work_request_in_filesystem(state_dir, request_id)
-            _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                request_id,
-                state="running",
-            )
-
-            close_status, closed_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=close_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue-close",
-                    "X-Hub-Signature-256": _github_webhook_signature(close_payload, secret),
-                },
-            )
-
-        self.assertEqual(create_status, 202)
-        self.assertEqual(close_status, 202)
-        self.assertEqual(closed_response["records"]["state"], "done")
-        self.assertEqual(
-            closed_response["result"]["request"]["finished_at"],
-            "2026-05-06T16:20:00Z",
-        )
-        self.assertEqual(
-            closed_response["result"]["request"]["result_summary"],
-            "Source issue closed (completed): https://github.com/cbusillo/code/issues/123",
-        )
-
-    def test_every_code_pull_request_close_marks_linked_request_done(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        policy = LaunchplaneAuthzPolicy.model_validate(
-            {
-                "github_actions": [
-                    {
-                        "repository": "cbusillo/launchplane",
-                        "workflow_refs": [
-                            "cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main"
-                        ],
-                        "event_names": ["workflow_dispatch"],
-                        "products": ["launchplane"],
-                        "contexts": ["launchplane"],
-                        "actions": [
-                            "every_code_work_request.read",
-                            "every_code_work_request.claim",
-                            "every_code_work_request.update",
-                        ],
-                    }
-                ]
-            }
-        )
-        identity = _identity(
-            repository="cbusillo/launchplane",
-            workflow_ref="cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main",
-            event_name="workflow_dispatch",
-        )
-        issue_payload = _every_code_github_issue_labeled_payload()
-        pr_payload = _every_code_github_pull_request_closed_payload()
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(identity),
-                authz_policy=policy,
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            create_status, create_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = str(create_payload["records"]["request_id"])
-            _claim_every_code_work_request_in_filesystem(state_dir, request_id)
-            _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                request_id,
-                state="running",
-                result_pr_url="https://github.com/cbusillo/code/pull/26",
-            )
-
-            close_status, close_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=pr_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "pull_request",
-                    "X-GitHub-Delivery": "delivery-pr-close",
-                    "X-Hub-Signature-256": _github_webhook_signature(pr_payload, secret),
-                },
-            )
-
-        self.assertEqual(create_status, 202)
-        self.assertEqual(close_status, 202)
-        self.assertEqual(close_payload["records"]["state"], "done")
-        self.assertEqual(close_payload["result"]["request"]["finished_at"], "2026-05-06T16:20:00Z")
-        self.assertEqual(close_payload["result"]["request"]["error_message"], "")
-        self.assertEqual(
-            close_payload["result"]["request"]["result_summary"],
-            "Linked pull request merged: https://github.com/cbusillo/code/pull/26",
-        )
-
-    def test_every_code_pull_request_close_blocks_unmerged_linked_request(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        policy = LaunchplaneAuthzPolicy.model_validate(
-            {
-                "github_actions": [
-                    {
-                        "repository": "cbusillo/launchplane",
-                        "workflow_refs": [
-                            "cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main"
-                        ],
-                        "event_names": ["workflow_dispatch"],
-                        "products": ["launchplane"],
-                        "contexts": ["launchplane"],
-                        "actions": [
-                            "every_code_work_request.read",
-                            "every_code_work_request.claim",
-                            "every_code_work_request.update",
-                        ],
-                    }
-                ]
-            }
-        )
-        identity = _identity(
-            repository="cbusillo/launchplane",
-            workflow_ref="cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main",
-            event_name="workflow_dispatch",
-        )
-        issue_payload = _every_code_github_issue_labeled_payload()
-        pr_payload = _every_code_github_pull_request_closed_payload(merged=False)
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {"LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret},
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(identity),
-                authz_policy=policy,
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            create_status, create_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = str(create_payload["records"]["request_id"])
-            _claim_every_code_work_request_in_filesystem(state_dir, request_id)
-            _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                request_id,
-                state="running",
-                result_pr_url="https://github.com/cbusillo/code/pull/26",
-            )
-
-            close_status, close_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=pr_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "pull_request",
-                    "X-GitHub-Delivery": "delivery-pr-close",
-                    "X-Hub-Signature-256": _github_webhook_signature(pr_payload, secret),
-                },
-            )
-
-        self.assertEqual(create_status, 202)
-        self.assertEqual(close_status, 202)
-        self.assertEqual(close_payload["records"]["state"], "blocked")
-        self.assertIn("closed without merge", close_payload["result"]["request"]["error_message"])
-
-    def test_every_code_pull_request_close_matches_issue_url_without_result_pr_url(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        policy = LaunchplaneAuthzPolicy.model_validate(
-            {
-                "github_actions": [
-                    {
-                        "repository": "cbusillo/launchplane",
-                        "workflow_refs": [
-                            "cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main"
-                        ],
-                        "event_names": ["workflow_dispatch"],
-                        "products": ["launchplane"],
-                        "contexts": ["launchplane"],
-                        "actions": [
-                            "every_code_work_request.read",
-                            "every_code_work_request.claim",
-                            "every_code_work_request.update",
-                        ],
-                    }
-                ]
-            }
-        )
-        identity = _identity(
-            repository="cbusillo/launchplane",
-            workflow_ref="cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main",
-            event_name="workflow_dispatch",
-        )
-        issue_payload = _every_code_github_issue_labeled_payload(issue_number=64)
-        pr_payload = _every_code_github_pull_request_closed_payload(
-            pr_number=71,
-            body="Closes #64",
-        )
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {"LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret},
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(identity),
-                authz_policy=policy,
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            create_status, create_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-feedback-close-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = str(create_payload["records"]["request_id"])
-            _claim_every_code_work_request_in_filesystem(state_dir, request_id)
-            _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                request_id,
-                state="running",
-            )
-            close_status, close_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=pr_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "pull_request",
-                    "X-GitHub-Delivery": "delivery-feedback-close-pr",
-                    "X-Hub-Signature-256": _github_webhook_signature(pr_payload, secret),
-                },
-            )
-
-        self.assertEqual(create_status, 202)
-        self.assertEqual(close_status, 202)
-        self.assertEqual(close_payload["records"]["state"], "done")
-        self.assertEqual(
-            close_payload["result"]["request"]["result_pr_url"],
-            "https://github.com/cbusillo/code/pull/71",
-        )
-
-    def test_every_code_pull_request_close_matches_all_linked_issue_urls(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        policy = LaunchplaneAuthzPolicy.model_validate(
-            {
-                "github_actions": [
-                    {
-                        "repository": "cbusillo/launchplane",
-                        "workflow_refs": [
-                            "cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main"
-                        ],
-                        "event_names": ["workflow_dispatch"],
-                        "products": ["launchplane"],
-                        "contexts": ["launchplane"],
-                        "actions": ["every_code_work_request.read"],
-                    }
-                ]
-            }
-        )
-        identity = _identity(
-            repository="cbusillo/launchplane",
-            workflow_ref="cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main",
-            event_name="workflow_dispatch",
-        )
-        pr_payload = _every_code_github_pull_request_closed_payload(
-            pr_number=71,
-            body="Closes #64, closes #65",
-        )
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(identity),
-                authz_policy=policy,
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            for issue_number in (64, 65):
-                issue_payload = _every_code_github_issue_labeled_payload(issue_number=issue_number)
-                _invoke_app(
-                    app,
-                    method="POST",
-                    path="/v1/every-code/github-webhook",
-                    payload=issue_payload,
-                    authorization="",
-                    headers={
-                        "X-GitHub-Event": "issues",
-                        "X-GitHub-Delivery": f"delivery-multi-close-{issue_number}",
-                        "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                    },
-                )
-
-            close_status, close_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=pr_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "pull_request",
-                    "X-GitHub-Delivery": "delivery-multi-close-pr",
-                    "X-Hub-Signature-256": _github_webhook_signature(pr_payload, secret),
-                },
-            )
-            persisted_requests = FilesystemRecordStore(
-                state_dir
-            ).list_every_code_work_request_records()
-
-        self.assertEqual(close_status, 202)
-        self.assertEqual(close_payload["result"]["closed_count"], 2)
-        closed_requests = close_payload["result"]["requests"]
-        self.assertEqual({request["issue_number"] for request in closed_requests}, {64, 65})
-        self.assertTrue(all(request["state"] == "done" for request in closed_requests))
-        self.assertTrue(
-            all(
-                request["claimed_by_host"] == "github-pull-request-close"
-                for request in closed_requests
-            )
-        )
-        self.assertEqual(
-            {request.state for request in persisted_requests},
-            {"done"},
-        )
-
-    def test_every_code_pull_request_close_does_not_match_issue_by_pr_number(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        policy = LaunchplaneAuthzPolicy.model_validate(
-            {
-                "github_actions": [
-                    {
-                        "repository": "cbusillo/launchplane",
-                        "workflow_refs": [
-                            "cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main"
-                        ],
-                        "event_names": ["workflow_dispatch"],
-                        "products": ["launchplane"],
-                        "contexts": ["launchplane"],
-                        "actions": [
-                            "every_code_work_request.read",
-                            "every_code_work_request.claim",
-                            "every_code_work_request.update",
-                        ],
-                    }
-                ]
-            }
-        )
-        identity = _identity(
-            repository="cbusillo/launchplane",
-            workflow_ref="cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main",
-            event_name="workflow_dispatch",
-        )
-        issue_payload = _every_code_github_issue_labeled_payload(issue_number=26)
-        pr_payload = _every_code_github_pull_request_closed_payload(pr_number=26)
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {"LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret},
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(identity),
-                authz_policy=policy,
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            create_status, create_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-pr-number-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = str(create_payload["records"]["request_id"])
-            _claim_every_code_work_request_in_filesystem(state_dir, request_id)
-            _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                request_id,
-                state="running",
-            )
-            close_status, close_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=pr_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "pull_request",
-                    "X-GitHub-Delivery": "delivery-pr-number-close",
-                    "X-Hub-Signature-256": _github_webhook_signature(pr_payload, secret),
-                },
-            )
-
-        self.assertEqual(create_status, 202)
-        self.assertEqual(close_status, 202)
-        self.assertTrue(close_payload["skipped"])
-        self.assertEqual(close_payload["reason"], "linked_every_code_request_not_found")
-
-    def test_every_code_pull_request_close_pages_beyond_newest_requests(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        policy = LaunchplaneAuthzPolicy.model_validate(
-            {
-                "github_actions": [
-                    {
-                        "repository": "cbusillo/launchplane",
-                        "workflow_refs": [
-                            "cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main"
-                        ],
-                        "event_names": ["workflow_dispatch"],
-                        "products": ["launchplane"],
-                        "contexts": ["launchplane"],
-                        "actions": [
-                            "every_code_work_request.read",
-                            "every_code_work_request.claim",
-                            "every_code_work_request.update",
-                        ],
-                    }
-                ]
-            }
-        )
-        identity = _identity(
-            repository="cbusillo/launchplane",
-            workflow_ref="cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main",
-            event_name="workflow_dispatch",
-        )
-        target_pr_payload = _every_code_github_pull_request_closed_payload(pr_number=26)
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {"LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret},
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(identity),
-                authz_policy=policy,
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            for issue_number in range(1, 103):
-                issue_payload = _every_code_github_issue_labeled_payload(issue_number=issue_number)
-                _create_status, create_payload = _invoke_app(
-                    app,
-                    method="POST",
-                    path="/v1/every-code/github-webhook",
-                    payload=issue_payload,
-                    authorization="",
-                    headers={
-                        "X-GitHub-Event": "issues",
-                        "X-GitHub-Delivery": f"delivery-page-issue-{issue_number}",
-                        "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                    },
-                )
-                request_id = str(create_payload["records"]["request_id"])
-                _claim_every_code_work_request_in_filesystem(state_dir, request_id)
-                _update_every_code_work_request_status_in_filesystem(
-                    state_dir,
-                    request_id,
-                    state="running",
-                    result_pr_url=f"https://github.com/cbusillo/code/pull/{issue_number}",
-                )
-
-            close_status, close_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=target_pr_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "pull_request",
-                    "X-GitHub-Delivery": "delivery-page-pr-close",
-                    "X-Hub-Signature-256": _github_webhook_signature(target_pr_payload, secret),
-                },
-            )
-
-        self.assertEqual(close_status, 202)
-        self.assertEqual(close_payload["records"]["state"], "done")
-        self.assertEqual(
-            close_payload["result"]["request"]["result_pr_url"],
-            "https://github.com/cbusillo/code/pull/26",
-        )
-
-    def test_every_code_pull_request_close_ignores_unlinked_pr(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        identity = _identity(
-            repository="cbusillo/launchplane",
-            workflow_ref="cbusillo/launchplane/.github/workflows/every-code-worker.yml@refs/heads/main",
-            event_name="workflow_dispatch",
-        )
-        pr_payload = _every_code_github_pull_request_closed_payload(pr_number=999)
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {"LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret},
-            ),
-        ):
-            app = create_every_code_github_webhook_app(
-                state_dir=Path(temporary_directory_name) / "state",
-                verifier=_StubVerifier(identity),
-                authz_policy=LaunchplaneAuthzPolicy(),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=pr_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "pull_request",
-                    "X-GitHub-Delivery": "delivery-pr-close",
-                    "X-Hub-Signature-256": _github_webhook_signature(pr_payload, secret),
-                },
-            )
-
-        self.assertEqual(status_code, 202)
-        self.assertTrue(payload["skipped"])
-        self.assertEqual(payload["reason"], "linked_every_code_request_not_found")
-
-    def test_every_code_pr_comment_webhook_records_feedback(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        issue_payload = _every_code_github_issue_labeled_payload()
-        comment_payload = _every_code_github_pr_comment_payload()
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy(),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            issue_status, issue_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = issue_response["records"]["request_id"]
-            claim_status, claim_payload = _claim_every_code_work_request_in_filesystem(
-                state_dir,
-                str(request_id),
-            )
-            status_status, _status_payload = _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                str(request_id),
-                state="done",
-                result_pr_url="https://github.com/cbusillo/code/pull/26",
-                result_summary="Opened PR.",
-                updated_at="2026-05-06T16:00:00Z",
-            )
-            feedback_status, feedback_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=comment_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issue_comment",
-                    "X-GitHub-Delivery": "delivery-comment",
-                    "X-Hub-Signature-256": _github_webhook_signature(comment_payload, secret),
-                },
-            )
-
-        self.assertEqual(issue_status, 202)
-        self.assertEqual(claim_status, 202)
-        self.assertEqual(status_status, 202)
-        self.assertEqual(claim_payload["result"]["request"]["state"], "claimed")
-        self.assertEqual(feedback_status, 202)
-        self.assertEqual(feedback_response["records"]["request_id"], request_id)
-        feedback = feedback_response["result"]["feedback"]
-        self.assertEqual(feedback["request_id"], request_id)
-        self.assertEqual(feedback["feedback_kind"], "issue_comment")
-        self.assertEqual(feedback["body"], "Please tighten this wording before merge.")
-
-    def test_every_code_preview_ok_comment_marks_pr_ready_to_merge(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        issue_payload = _every_code_github_issue_labeled_payload(
-            repository="cbusillo/sellyouroutboard",
-            issue_number=82,
-        )
-        comment_payload = _every_code_github_issue_comment_payload(
-            repository="cbusillo/sellyouroutboard",
-            issue_number=82,
-            body="/preview ok",
-        )
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-            patch(
-                "control_plane.workflows.preview_pr_feedback.resolve_launchplane_github_token",
-                return_value="github-token",
-            ),
-            patch(
-                "control_plane.workflows.preview_pr_feedback.github_api_request",
-                side_effect=[
-                    [{"name": "preview-approved"}],
-                    {},
-                    {},
-                    [{"name": "ready-to-merge"}],
-                    {"owner": {"login": "cbusillo", "type": "User"}},
-                    {"assignees": [{"login": "cbusillo"}]},
-                ],
-            ) as github_request,
-            patch(
-                "control_plane.workflows.preview_pr_feedback.find_github_issue_comment_by_marker",
-                return_value=None,
-            ),
-            patch(
-                "control_plane.workflows.preview_pr_feedback.create_github_issue_comment",
-                return_value={"id": 987},
-            ) as create_comment,
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy(),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            issue_status, issue_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = issue_response["records"]["request_id"]
-            claim_status, _claim_payload = _claim_every_code_work_request_in_filesystem(
-                state_dir,
-                str(request_id),
-            )
-            status_status, _status_payload = _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                str(request_id),
-                state="done",
-                result_pr_url="https://github.com/cbusillo/sellyouroutboard/pull/88",
-                result_summary="Opened PR.",
-                updated_at="2026-05-07T12:40:00Z",
-            )
-            ok_status, ok_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=comment_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issue_comment",
-                    "X-GitHub-Delivery": "delivery-preview-ok",
-                    "X-Hub-Signature-256": _github_webhook_signature(comment_payload, secret),
-                },
-            )
-
-        self.assertEqual(issue_status, 202)
-        self.assertEqual(claim_status, 202)
-        self.assertEqual(status_status, 202)
-        self.assertEqual(ok_status, 202, ok_response)
-        preview_validation = ok_response["result"]["preview_validation"]
-        self.assertEqual(preview_validation["command"], "ok")
-        self.assertEqual(preview_validation["merge_owner"], "cbusillo")
-        self.assertEqual(
-            github_request.call_args_list[3].kwargs["body"], {"labels": ["ready-to-merge"]}
-        )
-        self.assertEqual(
-            github_request.call_args_list[5].kwargs["body"], {"assignees": ["cbusillo"]}
-        )
-        create_comment.assert_called_once()
-        self.assertIn("@cbusillo", create_comment.call_args.kwargs["body"])
-
     def test_preview_pr_feedback_hydrates_ready_url_from_preview_record(self) -> None:
         with (
             TemporaryDirectory() as temporary_directory_name,
@@ -3384,1264 +2313,6 @@ class LaunchplaneServiceTests(unittest.TestCase):
         )
         create_comment.assert_called_once()
 
-    def test_every_code_preview_validation_failure_returns_generic_webhook_response(
-        self,
-    ) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        comment_payload = _every_code_github_issue_comment_payload(
-            repository="cbusillo/sellyouroutboard",
-            issue_number=82,
-            body="/preview ok",
-        )
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {"LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret},
-            ),
-            patch(
-                "control_plane.service.resolve_launchplane_github_token",
-                side_effect=ClickException(
-                    "Traceback (most recent call last): secret token leaked"
-                ),
-            ),
-        ):
-            app = create_every_code_github_webhook_app(
-                state_dir=Path(temporary_directory_name) / "state",
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy(),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=comment_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issue_comment",
-                    "X-GitHub-Delivery": "delivery-preview-validation-failed",
-                    "X-Hub-Signature-256": _github_webhook_signature(comment_payload, secret),
-                },
-            )
-
-        self.assertEqual(status_code, 202)
-        self.assertEqual(payload["status"], "accepted")
-        self.assertIs(payload["skipped"], True)
-        self.assertEqual(payload["reason"], "preview_validation_failed")
-        self.assertNotIn("message", payload)
-        self.assertNotIn("Traceback", json.dumps(payload, sort_keys=True))
-
-    def test_every_code_preview_ok_allows_repo_owner_override(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        issue_payload = _every_code_github_issue_labeled_payload(
-            repository="cbusillo/sellyouroutboard",
-            issue_number=82,
-        )
-        comment_payload = _every_code_github_issue_comment_payload(
-            repository="cbusillo/sellyouroutboard",
-            issue_number=82,
-            issue_author="Mbanks89",
-            sender="cbusillo",
-            body="/preview ok",
-        )
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-            patch(
-                "control_plane.service.resolve_launchplane_github_token",
-                return_value="github-token",
-            ),
-            patch(
-                "control_plane.workflows.preview_pr_feedback.github_api_request",
-                side_effect=[
-                    [{"name": "preview-approved"}],
-                    {},
-                    {},
-                    [{"name": "ready-to-merge"}],
-                    {"owner": {"login": "cbusillo", "type": "User"}},
-                    {"assignees": [{"login": "cbusillo"}]},
-                ],
-            ),
-            patch(
-                "control_plane.workflows.preview_pr_feedback.find_github_issue_comment_by_marker",
-                return_value=None,
-            ),
-            patch(
-                "control_plane.workflows.preview_pr_feedback.create_github_issue_comment",
-                return_value={"id": 987},
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy(),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            issue_status, issue_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = issue_response["records"]["request_id"]
-            _claim_every_code_work_request_in_filesystem(state_dir, str(request_id))
-            _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                str(request_id),
-                state="done",
-                result_pr_url="https://github.com/cbusillo/sellyouroutboard/pull/88",
-                result_summary="Opened PR.",
-                updated_at="2026-05-07T12:40:00Z",
-            )
-            ok_status, ok_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=comment_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issue_comment",
-                    "X-GitHub-Delivery": "delivery-preview-owner-ok",
-                    "X-Hub-Signature-256": _github_webhook_signature(comment_payload, secret),
-                },
-            )
-
-        self.assertEqual(issue_status, 202)
-        self.assertEqual(ok_status, 202, ok_response)
-        self.assertEqual(ok_response["result"]["preview_validation"]["command"], "ok")
-
-    def test_every_code_preview_comment_skips_untrusted_actor(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        issue_payload = _every_code_github_issue_labeled_payload(
-            repository="cbusillo/sellyouroutboard",
-            issue_number=82,
-        )
-        comment_payload = _every_code_github_issue_comment_payload(
-            repository="cbusillo/sellyouroutboard",
-            issue_number=82,
-            issue_author="Mbanks89",
-            sender="random-user",
-            body="/preview ok",
-        )
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy(),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            _issue_status, _issue_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            ok_status, ok_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=comment_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issue_comment",
-                    "X-GitHub-Delivery": "delivery-preview-untrusted-ok",
-                    "X-Hub-Signature-256": _github_webhook_signature(comment_payload, secret),
-                },
-            )
-
-        self.assertEqual(ok_status, 202, ok_response)
-        self.assertTrue(ok_response["skipped"])
-        self.assertEqual(ok_response["reason"], "untrusted_actor")
-
-    def test_every_code_preview_changes_routes_feedback_to_session(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        issue_payload = _every_code_github_issue_labeled_payload(
-            repository="cbusillo/sellyouroutboard",
-            issue_number=82,
-        )
-        comment_payload = _every_code_github_issue_comment_payload(
-            repository="cbusillo/sellyouroutboard",
-            issue_number=82,
-            body="/preview changes The delete button still misses bulk uploads.",
-        )
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-            patch(
-                "control_plane.service.resolve_launchplane_github_token",
-                return_value="github-token",
-            ),
-            patch(
-                "control_plane.workflows.preview_pr_feedback.github_api_request",
-                side_effect=[
-                    [{"name": "preview-changes-requested"}],
-                    {},
-                    {},
-                    {},
-                ],
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy(),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            issue_status, issue_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = issue_response["records"]["request_id"]
-            claim_status, _claim_payload = _claim_every_code_work_request_in_filesystem(
-                state_dir,
-                str(request_id),
-            )
-            status_status, _status_payload = _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                str(request_id),
-                state="done",
-                result_pr_url="https://github.com/cbusillo/sellyouroutboard/pull/88",
-                result_summary="Opened PR.",
-                updated_at="2026-05-07T12:40:00Z",
-            )
-            changes_status, changes_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=comment_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issue_comment",
-                    "X-GitHub-Delivery": "delivery-preview-changes",
-                    "X-Hub-Signature-256": _github_webhook_signature(comment_payload, secret),
-                },
-            )
-
-        self.assertEqual(issue_status, 202)
-        self.assertEqual(claim_status, 202)
-        self.assertEqual(status_status, 202)
-        self.assertEqual(changes_status, 202, changes_response)
-        preview_validation = changes_response["result"]["preview_validation"]
-        self.assertEqual(preview_validation["command"], "changes")
-        feedback = preview_validation["feedback_id"]
-        self.assertIn("every-code-pr-feedback-cbusillo-sellyouroutboard-88", feedback)
-
-    def test_every_code_pr_comment_webhook_matches_linked_issue(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        issue_payload = _every_code_github_issue_labeled_payload(issue_number=67)
-        comment_payload = _every_code_github_pr_comment_payload(
-            pr_number=75,
-            issue_body="Closes #67",
-        )
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy(),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            issue_status, issue_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = issue_response["records"]["request_id"]
-            claim_status, _claim_payload = _claim_every_code_work_request_in_filesystem(
-                state_dir,
-                str(request_id),
-            )
-            running_status, _running_payload = _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                str(request_id),
-                state="running",
-                result_summary="Visible tmux session.",
-                updated_at="2026-05-06T16:00:00Z",
-            )
-            feedback_status, feedback_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=comment_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issue_comment",
-                    "X-GitHub-Delivery": "delivery-comment",
-                    "X-Hub-Signature-256": _github_webhook_signature(comment_payload, secret),
-                },
-            )
-
-        self.assertEqual(issue_status, 202)
-        self.assertEqual(claim_status, 202)
-        self.assertEqual(running_status, 202)
-        self.assertIn("records", feedback_response)
-        self.assertEqual(feedback_status, 202, feedback_response)
-        self.assertEqual(feedback_response["records"]["request_id"], request_id)
-        feedback = feedback_response["result"]["feedback"]
-        self.assertEqual(feedback["request_id"], request_id)
-        self.assertEqual(feedback["pr_number"], 75)
-        self.assertEqual(feedback["pr_url"], "https://github.com/cbusillo/code/pull/75")
-
-    def test_every_code_pr_comment_webhook_allows_configured_manager(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        issue_payload = _every_code_github_issue_labeled_payload(
-            repository="cbusillo/sellyouroutboard",
-            issue_number=67,
-        )
-        comment_payload = _every_code_github_pr_comment_payload(
-            repository="cbusillo/sellyouroutboard",
-            pr_number=75,
-            issue_body="Closes #67",
-            sender="Mbanks89",
-        )
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            tempfile.TemporaryDirectory() as home_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "HOME": home_directory_name,
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            _write_github_planning_config(
-                Path(home_directory_name),
-                repo_managers={"cbusillo/sellyouroutboard": "@Mbanks89"},
-            )
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy(),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            issue_status, issue_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = issue_response["records"]["request_id"]
-            _claim_every_code_work_request_in_filesystem(state_dir, str(request_id))
-            _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                str(request_id),
-                state="running",
-                result_summary="Visible tmux session.",
-                updated_at="2026-05-06T16:00:00Z",
-            )
-            feedback_status, feedback_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=comment_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issue_comment",
-                    "X-GitHub-Delivery": "delivery-comment-manager",
-                    "X-Hub-Signature-256": _github_webhook_signature(comment_payload, secret),
-                },
-            )
-
-        self.assertEqual(issue_status, 202)
-        self.assertEqual(feedback_status, 202, feedback_response)
-        self.assertEqual(feedback_response["records"]["request_id"], request_id)
-        self.assertEqual(feedback_response["result"]["feedback"]["actor"], "Mbanks89")
-
-    def test_every_code_pr_comment_webhook_uses_second_planning_config_path(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        issue_payload = _every_code_github_issue_labeled_payload(
-            repository="cbusillo/sellyouroutboard",
-            issue_number=67,
-        )
-        comment_payload = _every_code_github_pr_comment_payload(
-            repository="cbusillo/sellyouroutboard",
-            pr_number=75,
-            issue_body="Closes #67",
-            sender="Mbanks89",
-        )
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            tempfile.TemporaryDirectory() as home_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "HOME": home_directory_name,
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            _write_github_planning_config(Path(home_directory_name), repo_managers={})
-            _write_github_planning_config(
-                Path(home_directory_name),
-                path=".codex/github-planning.json",
-                repo_managers={"cbusillo/sellyouroutboard": "@Mbanks89"},
-            )
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy(),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            issue_status, issue_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue-second-config",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = issue_response["records"]["request_id"]
-            _claim_every_code_work_request_in_filesystem(state_dir, str(request_id))
-            _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                str(request_id),
-                state="done",
-                result_pr_url="https://github.com/cbusillo/sellyouroutboard/pull/75",
-                result_summary="Opened PR.",
-                updated_at="2026-05-07T12:40:00Z",
-            )
-            feedback_status, feedback_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=comment_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issue_comment",
-                    "X-GitHub-Delivery": "delivery-comment-second-config",
-                    "X-Hub-Signature-256": _github_webhook_signature(comment_payload, secret),
-                },
-            )
-
-        self.assertEqual(issue_status, 202)
-        self.assertEqual(feedback_status, 202, feedback_response)
-        self.assertEqual(feedback_response["records"]["request_id"], request_id)
-        self.assertEqual(feedback_response["result"]["feedback"]["actor"], "Mbanks89")
-
-    def test_every_code_pr_comment_webhook_skips_untrusted_actor(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        issue_payload = _every_code_github_issue_labeled_payload(issue_number=67)
-        comment_payload = _every_code_github_pr_comment_payload(
-            pr_number=75,
-            issue_body="Closes #67",
-            sender="random-user",
-        )
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            tempfile.TemporaryDirectory() as home_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "HOME": home_directory_name,
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            _write_github_planning_config(Path(home_directory_name), repo_managers={})
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy(),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            issue_status, issue_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = issue_response["records"]["request_id"]
-            feedback_status, feedback_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=comment_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issue_comment",
-                    "X-GitHub-Delivery": "delivery-comment-untrusted",
-                    "X-Hub-Signature-256": _github_webhook_signature(comment_payload, secret),
-                },
-            )
-            feedback_records = FilesystemRecordStore(state_dir).list_every_code_pr_feedback_records(
-                request_id=request_id
-            )
-
-        self.assertEqual(issue_status, 202)
-        self.assertEqual(feedback_status, 202, feedback_response)
-        self.assertTrue(feedback_response["skipped"])
-        self.assertEqual(feedback_response["reason"], "untrusted_actor")
-        self.assertEqual(feedback_records, ())
-
-    def test_every_code_pr_comment_webhook_ignores_bot_feedback(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        issue_payload = _every_code_github_issue_labeled_payload(issue_number=67)
-        comment_payload = _every_code_github_pr_comment_payload(
-            pr_number=75,
-            issue_body="Closes #67",
-            body="Odoo preview refresh started for PR #75.",
-            sender="github-actions[bot]",
-            sender_type="Bot",
-        )
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy(),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            issue_status, issue_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = issue_response["records"]["request_id"]
-            _claim_status, _claim_payload = _claim_every_code_work_request_in_filesystem(
-                state_dir,
-                str(request_id),
-            )
-            _running_status, _running_payload = (
-                _update_every_code_work_request_status_in_filesystem(
-                    state_dir,
-                    str(request_id),
-                    state="running",
-                    result_summary="Visible tmux session.",
-                    updated_at="2026-05-06T16:00:00Z",
-                )
-            )
-            feedback_status, feedback_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=comment_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issue_comment",
-                    "X-GitHub-Delivery": "delivery-comment-bot",
-                    "X-Hub-Signature-256": _github_webhook_signature(comment_payload, secret),
-                },
-            )
-            feedback_records = FilesystemRecordStore(state_dir).list_every_code_pr_feedback_records(
-                request_id=request_id
-            )
-
-        self.assertEqual(issue_status, 202)
-        self.assertEqual(feedback_status, 202, feedback_response)
-        self.assertTrue(feedback_response["skipped"])
-        self.assertEqual(feedback_response["reason"], "automation_actor")
-        self.assertEqual(feedback_records, ())
-
-    def test_every_code_pr_comment_webhook_dedupes_feedback(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        issue_payload = _every_code_github_issue_labeled_payload()
-        comment_payload = _every_code_github_pr_comment_payload(comment_id=2002)
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy(),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            issue_status, issue_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = issue_response["records"]["request_id"]
-            _claim_every_code_work_request_in_filesystem(state_dir, str(request_id))
-            _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                str(request_id),
-                state="done",
-                result_pr_url="https://github.com/cbusillo/code/pull/26",
-                result_summary="Opened PR.",
-                updated_at="2026-05-06T16:00:00Z",
-            )
-            first_status, first_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=comment_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issue_comment",
-                    "X-GitHub-Delivery": "delivery-comment",
-                    "X-Hub-Signature-256": _github_webhook_signature(comment_payload, secret),
-                },
-            )
-            second_status, second_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=comment_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issue_comment",
-                    "X-GitHub-Delivery": "delivery-comment",
-                    "X-Hub-Signature-256": _github_webhook_signature(comment_payload, secret),
-                },
-            )
-
-        self.assertEqual(issue_status, 202)
-        self.assertEqual(first_status, 202)
-        self.assertEqual(second_status, 202)
-        self.assertEqual(
-            first_response["result"]["feedback"]["feedback_id"],
-            second_response["result"]["feedback_id"],
-        )
-        self.assertTrue(second_response["deduped"])
-
-    def test_every_code_github_webhook_rejects_invalid_signature(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        webhook_payload = _every_code_github_issue_labeled_payload()
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {"LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret},
-            ),
-        ):
-            app = create_every_code_github_webhook_app(
-                state_dir=Path(temporary_directory_name) / "state",
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy.model_validate({"github_actions": []}),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=webhook_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-1",
-                    "X-Hub-Signature-256": "sha256=invalid",
-                },
-            )
-
-        self.assertEqual(status_code, 401)
-        self.assertEqual(payload["error"]["code"], "webhook_signature_invalid")
-
-    def test_every_code_github_webhook_rejects_invalid_json_payload(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        body_bytes = b'{"action":'
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {"LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret},
-            ),
-        ):
-            app = create_every_code_github_webhook_app(
-                state_dir=Path(temporary_directory_name) / "state",
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy.model_validate({"github_actions": []}),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            status_code, _headers, response_body = _invoke_http_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                authorization="",
-                body_bytes=body_bytes,
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-invalid-json",
-                    "X-Hub-Signature-256": _github_webhook_body_signature(
-                        body_bytes,
-                        secret,
-                    ),
-                },
-            )
-            payload = json.loads(response_body.decode("utf-8"))
-
-        self.assertEqual(status_code, 400)
-        self.assertEqual(payload["status"], "rejected")
-        self.assertEqual(payload["error"]["code"], "invalid_request")
-        self.assertEqual(payload["error"]["message"], "GitHub webhook payload is invalid.")
-
-    def test_every_code_github_webhook_rejects_malformed_issue_payload(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        webhook_payload = _every_code_github_issue_labeled_payload()
-        webhook_payload["repository"] = {}
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {"LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret},
-            ),
-        ):
-            app = create_every_code_github_webhook_app(
-                state_dir=Path(temporary_directory_name) / "state",
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy.model_validate({"github_actions": []}),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=webhook_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-malformed-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(webhook_payload, secret),
-                },
-            )
-
-        self.assertEqual(status_code, 400)
-        self.assertEqual(payload["status"], "rejected")
-        self.assertEqual(payload["error"]["code"], "invalid_request")
-        self.assertEqual(payload["error"]["message"], "GitHub webhook payload is invalid.")
-
-    def test_every_code_github_webhook_rejects_bool_issue_number(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        webhook_payload = _every_code_github_issue_labeled_payload()
-        issue = cast(dict[str, object], webhook_payload["issue"])
-        issue["number"] = True
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {"LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret},
-            ),
-        ):
-            app = create_every_code_github_webhook_app(
-                state_dir=Path(temporary_directory_name) / "state",
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy.model_validate({"github_actions": []}),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=webhook_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-bool-issue-number",
-                    "X-Hub-Signature-256": _github_webhook_signature(webhook_payload, secret),
-                },
-            )
-
-        self.assertEqual(status_code, 400)
-        self.assertEqual(payload["status"], "rejected")
-        self.assertEqual(payload["error"]["code"], "invalid_request")
-        self.assertEqual(payload["error"]["message"], "GitHub webhook payload is invalid.")
-
-    def test_every_code_github_webhook_rejects_malformed_repository_name(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        webhook_payload = _every_code_github_issue_labeled_payload()
-        repository = cast(dict[str, object], webhook_payload["repository"])
-        repository["full_name"] = " cbusillo/code"
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {"LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret},
-            ),
-        ):
-            app = create_every_code_github_webhook_app(
-                state_dir=Path(temporary_directory_name) / "state",
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy.model_validate({"github_actions": []}),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=webhook_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-malformed-repository",
-                    "X-Hub-Signature-256": _github_webhook_signature(webhook_payload, secret),
-                },
-            )
-
-        self.assertEqual(status_code, 400)
-        self.assertEqual(payload["status"], "rejected")
-        self.assertEqual(payload["error"]["code"], "invalid_request")
-        self.assertEqual(payload["error"]["message"], "GitHub webhook payload is invalid.")
-
-    def test_every_code_github_webhook_rejects_malformed_pull_request_repository(
-        self,
-    ) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        webhook_payload = _every_code_github_pull_request_closed_payload()
-        repository = cast(dict[str, object], webhook_payload["repository"])
-        repository["full_name"] = " cbusillo/code"
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {"LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret},
-            ),
-        ):
-            app = create_every_code_github_webhook_app(
-                state_dir=Path(temporary_directory_name) / "state",
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy.model_validate({"github_actions": []}),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=webhook_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "pull_request",
-                    "X-GitHub-Delivery": "delivery-malformed-pr-repository",
-                    "X-Hub-Signature-256": _github_webhook_signature(webhook_payload, secret),
-                },
-            )
-
-        self.assertEqual(status_code, 400)
-        self.assertEqual(payload["status"], "rejected")
-        self.assertEqual(payload["error"]["code"], "invalid_request")
-        self.assertEqual(payload["error"]["message"], "GitHub webhook payload is invalid.")
-
-    def test_every_code_github_webhook_rejects_malformed_pull_request_url(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        webhook_payload = _every_code_github_pull_request_closed_payload()
-        pull_request = cast(dict[str, object], webhook_payload["pull_request"])
-        pull_request["html_url"] = ""
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {"LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret},
-            ),
-        ):
-            app = create_every_code_github_webhook_app(
-                state_dir=Path(temporary_directory_name) / "state",
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy.model_validate({"github_actions": []}),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=webhook_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "pull_request",
-                    "X-GitHub-Delivery": "delivery-malformed-pr-url",
-                    "X-Hub-Signature-256": _github_webhook_signature(webhook_payload, secret),
-                },
-            )
-
-        self.assertEqual(status_code, 400)
-        self.assertEqual(payload["status"], "rejected")
-        self.assertEqual(payload["error"]["code"], "invalid_request")
-        self.assertEqual(payload["error"]["message"], "GitHub webhook payload is invalid.")
-
-    def test_every_code_github_webhook_rejects_malformed_feedback_payload(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        issue_payload = _every_code_github_issue_labeled_payload()
-        comment_payload = _every_code_github_pr_comment_payload()
-        comment = cast(dict[str, object], comment_payload["comment"])
-        comment.pop("node_id")
-        comment.pop("id")
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy(),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            issue_status, issue_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = str(issue_response["records"]["request_id"])
-            _claim_every_code_work_request_in_filesystem(state_dir, request_id)
-            _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                request_id,
-                state="done",
-                result_pr_url="https://github.com/cbusillo/code/pull/26",
-                result_summary="Opened PR.",
-                updated_at="2026-05-06T16:00:00Z",
-            )
-            feedback_status, feedback_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=comment_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issue_comment",
-                    "X-GitHub-Delivery": "delivery-malformed-feedback",
-                    "X-Hub-Signature-256": _github_webhook_signature(comment_payload, secret),
-                },
-            )
-            feedback_records = FilesystemRecordStore(state_dir).list_every_code_pr_feedback_records(
-                request_id=request_id
-            )
-
-        self.assertEqual(issue_status, 202)
-        self.assertEqual(feedback_status, 400)
-        self.assertEqual(feedback_response["status"], "rejected")
-        self.assertEqual(feedback_response["error"]["code"], "invalid_request")
-        self.assertEqual(
-            feedback_response["error"]["message"], "GitHub webhook payload is invalid."
-        )
-        self.assertEqual(feedback_records, ())
-
-    def test_every_code_github_webhook_rejects_malformed_feedback_repository(
-        self,
-    ) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        issue_payload = _every_code_github_issue_labeled_payload()
-        comment_payload = _every_code_github_pr_comment_payload()
-        repository = cast(dict[str, object], comment_payload["repository"])
-        repository["full_name"] = " cbusillo/code"
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy(),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            issue_status, issue_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = str(issue_response["records"]["request_id"])
-            _claim_every_code_work_request_in_filesystem(state_dir, request_id)
-            _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                request_id,
-                state="done",
-                result_pr_url="https://github.com/cbusillo/code/pull/26",
-                result_summary="Opened PR.",
-                updated_at="2026-05-06T16:00:00Z",
-            )
-            feedback_status, feedback_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=comment_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issue_comment",
-                    "X-GitHub-Delivery": "delivery-malformed-feedback-repository",
-                    "X-Hub-Signature-256": _github_webhook_signature(comment_payload, secret),
-                },
-            )
-            feedback_records = FilesystemRecordStore(state_dir).list_every_code_pr_feedback_records(
-                request_id=request_id
-            )
-
-        self.assertEqual(issue_status, 202)
-        self.assertEqual(feedback_status, 400)
-        self.assertEqual(feedback_response["status"], "rejected")
-        self.assertEqual(feedback_response["error"]["code"], "invalid_request")
-        self.assertEqual(
-            feedback_response["error"]["message"], "GitHub webhook payload is invalid."
-        )
-        self.assertEqual(feedback_records, ())
-
-    def test_every_code_github_webhook_rejects_slug_unsafe_feedback_identity(
-        self,
-    ) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        issue_payload = _every_code_github_issue_labeled_payload()
-        comment_payload = _every_code_github_pr_comment_payload()
-        comment = cast(dict[str, object], comment_payload["comment"])
-        comment["node_id"] = "---"
-        comment.pop("id")
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {
-                    "LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret,
-                    "LAUNCHPLANE_EVERY_CODE_WORKER_TOKEN": "dev-worker-token",
-                },
-            ),
-        ):
-            state_dir = Path(temporary_directory_name) / "state"
-            app = create_every_code_github_webhook_app(
-                state_dir=state_dir,
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy(),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            issue_status, issue_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=issue_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-issue",
-                    "X-Hub-Signature-256": _github_webhook_signature(issue_payload, secret),
-                },
-            )
-            request_id = str(issue_response["records"]["request_id"])
-            _claim_every_code_work_request_in_filesystem(state_dir, request_id)
-            _update_every_code_work_request_status_in_filesystem(
-                state_dir,
-                request_id,
-                state="done",
-                result_pr_url="https://github.com/cbusillo/code/pull/26",
-                result_summary="Opened PR.",
-                updated_at="2026-05-06T16:00:00Z",
-            )
-            feedback_status, feedback_response = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=comment_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issue_comment",
-                    "X-GitHub-Delivery": "delivery-slug-unsafe-feedback",
-                    "X-Hub-Signature-256": _github_webhook_signature(comment_payload, secret),
-                },
-            )
-            feedback_records = FilesystemRecordStore(state_dir).list_every_code_pr_feedback_records(
-                request_id=request_id
-            )
-
-        self.assertEqual(issue_status, 202)
-        self.assertEqual(feedback_status, 400)
-        self.assertEqual(feedback_response["status"], "rejected")
-        self.assertEqual(feedback_response["error"]["code"], "invalid_request")
-        self.assertEqual(
-            feedback_response["error"]["message"], "GitHub webhook payload is invalid."
-        )
-        self.assertEqual(feedback_records, ())
-
-    def test_every_code_github_webhook_ignores_other_labels(self) -> None:
-        secret = "launchplane-every-code-webhook-secret"
-        webhook_payload = _every_code_github_issue_labeled_payload(label="bug")
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {"LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": secret},
-            ),
-        ):
-            app = create_every_code_github_webhook_app(
-                state_dir=Path(temporary_directory_name) / "state",
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy.model_validate({"github_actions": []}),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=webhook_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-1",
-                    "X-Hub-Signature-256": _github_webhook_signature(webhook_payload, secret),
-                },
-            )
-
-        self.assertEqual(status_code, 202)
-        self.assertTrue(payload["skipped"])
-        self.assertEqual(payload["reason"], "label_not_matched")
-
-    def test_every_code_github_webhook_requires_configured_secret(self) -> None:
-        webhook_payload = _every_code_github_issue_labeled_payload()
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {"LAUNCHPLANE_EVERY_CODE_GITHUB_WEBHOOK_SECRET": ""},
-            ),
-        ):
-            app = create_every_code_github_webhook_app(
-                state_dir=Path(temporary_directory_name) / "state",
-                verifier=_StubVerifier(_identity()),
-                authz_policy=LaunchplaneAuthzPolicy.model_validate({"github_actions": []}),
-                control_plane_root_path=Path(temporary_directory_name),
-            )
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/every-code/github-webhook",
-                payload=webhook_payload,
-                authorization="",
-                headers={
-                    "X-GitHub-Event": "issues",
-                    "X-GitHub-Delivery": "delivery-1",
-                    "X-Hub-Signature-256": _github_webhook_signature(
-                        webhook_payload, "unused-secret"
-                    ),
-                },
-            )
-
-        self.assertEqual(status_code, 503)
-        self.assertEqual(payload["error"]["code"], "webhook_secret_not_configured")
-
     def test_service_serve_rejects_missing_database_url(self) -> None:
         runner = CliRunner()
         with TemporaryDirectory() as temporary_directory_name:
@@ -4711,163 +2382,6 @@ class LaunchplaneServiceTests(unittest.TestCase):
             payload["paths"],
         )
         self.assertNotIn('"examples"', json.dumps(payload, sort_keys=True))
-
-    def test_service_serve_runs_fastapi_app(self) -> None:
-        fastapi_response: Any | None = None
-        ui_response: Any | None = None
-        denied_config_response: Any | None = None
-        grant_response: Any | None = None
-        authorized_config_response: Any | None = None
-
-        def capture_uvicorn_run(app: Any, **_kwargs: object) -> None:
-            nonlocal fastapi_response, ui_response, denied_config_response
-            nonlocal grant_response, authorized_config_response
-            fastapi_response = asyncio.run(_http_get_for_service_test(app, "/openapi.json"))
-            ui_response = asyncio.run(_http_get_for_service_test(app, "/ui"))
-            denied_config_response = asyncio.run(
-                _http_request_for_service_test(
-                    app,
-                    method="GET",
-                    path="/v1/products/example-site/environments/prod/config-status",
-                    authorization="Bearer valid-token",
-                )
-            )
-            grant_response = asyncio.run(
-                _http_request_for_service_test(
-                    app,
-                    method="POST",
-                    path="/v1/authz-policies/github-actions/grants",
-                    authorization="Bearer valid-token",
-                    payload={
-                        "schema_version": 1,
-                        "product": "launchplane",
-                        "mode": "apply",
-                        "reason": "Grant FastAPI config-status read for runtime policy refresh test.",
-                        "related_issue": "cbusillo/launchplane#1323",
-                        "grant": {
-                            "repository": "cbusillo/launchplane",
-                            "workflow_refs": [
-                                "cbusillo/launchplane/.github/workflows/deploy-launchplane.yml@refs/heads/main"
-                            ],
-                            "event_names": ["workflow_dispatch"],
-                            "products": ["example-site"],
-                            "contexts": ["example-site"],
-                            "actions": ["product_environment.read"],
-                            "source_label": "test:fastapi-runtime-policy-refresh",
-                        },
-                    },
-                    headers={"Idempotency-Key": "authz-grant:fastapi-runtime-refresh"},
-                )
-            )
-            authorized_config_response = asyncio.run(
-                _http_request_for_service_test(
-                    app,
-                    method="GET",
-                    path="/v1/products/example-site/environments/prod/config-status",
-                    authorization="Bearer valid-token",
-                )
-            )
-
-        with TemporaryDirectory() as temporary_directory_name:
-            root = Path(temporary_directory_name)
-            ui_root = root / "control_plane" / "ui_static"
-            ui_root.mkdir(parents=True)
-            (ui_root / "index.html").write_text("<html>Launchplane UI</html>", encoding="utf-8")
-            policy_file = root / "policy.toml"
-            policy_file.write_text(
-                "\n".join(
-                    (
-                        "schema_version = 1",
-                        "",
-                        "[[github_actions]]",
-                        'repository = "cbusillo/launchplane"',
-                        "workflow_refs = [",
-                        '  "cbusillo/launchplane/.github/workflows/deploy-launchplane.yml@refs/heads/main",',
-                        "]",
-                        'event_names = ["workflow_dispatch"]',
-                        'products = ["launchplane"]',
-                        'contexts = ["launchplane"]',
-                        'actions = ["authz_policy_grant.write"]',
-                        "",
-                    )
-                ),
-                encoding="utf-8",
-            )
-            database_url = _sqlite_database_url(root / "launchplane.sqlite3")
-            store = PostgresRecordStore(database_url=database_url)
-            store.ensure_schema()
-            store.write_product_profile_record(
-                LaunchplaneProductProfileRecord.model_validate(_generic_site_profile_payload())
-            )
-            store.close()
-
-            with (
-                patch("uvicorn.run", side_effect=capture_uvicorn_run) as run_uvicorn,
-                patch("control_plane.service.control_plane_root", return_value=root),
-                patch(
-                    "control_plane.service_auth.GitHubOidcVerifier",
-                    return_value=_StubVerifier(
-                        _identity(
-                            repository="cbusillo/launchplane",
-                            workflow_ref=(
-                                "cbusillo/launchplane/.github/workflows/deploy-launchplane.yml@refs/heads/main"
-                            ),
-                            event_name="workflow_dispatch",
-                        )
-                    ),
-                ),
-            ):
-                control_plane_service.serve_launchplane_service(
-                    state_dir=root / "state",
-                    policy_file=policy_file,
-                    host="127.0.0.1",
-                    port=8080,
-                    audience="launchplane-service",
-                    database_url=database_url,
-                )
-
-            run_uvicorn.assert_called_once()
-
-        self.assertIsNotNone(fastapi_response)
-        self.assertIsNotNone(ui_response)
-        self.assertIsNotNone(denied_config_response)
-        self.assertIsNotNone(grant_response)
-        self.assertIsNotNone(authorized_config_response)
-        assert fastapi_response is not None
-        assert ui_response is not None
-        assert denied_config_response is not None
-        assert grant_response is not None
-        assert authorized_config_response is not None
-        self.assertEqual(fastapi_response.status_code, 200)
-        self.assertIn(
-            "/v1/products/{product}/environments/{environment}/config-status",
-            fastapi_response.json()["paths"],
-        )
-        self.assertEqual(ui_response.status_code, 200)
-        self.assertEqual(ui_response.headers["content-type"], "text/html")
-        self.assertIn(b"Launchplane UI", ui_response.content)
-        self.assertEqual(
-            denied_config_response.status_code,
-            403,
-            msg=denied_config_response.content.decode("utf-8"),
-        )
-        self.assertEqual(grant_response.status_code, 202)
-        self.assertEqual(grant_response.json()["result"]["changed"], True)
-        self.assertEqual(authorized_config_response.status_code, 200)
-        self.assertEqual(
-            authorized_config_response.json()["config_status"]["product"], "example-site"
-        )
-
-    def test_unregistered_descriptor_driver_route_fails_closed(self) -> None:
-        with patch(
-            "control_plane.drivers.native_routes.list_driver_descriptors",
-            return_value=(_fake_descriptor_dispatch_descriptor(),),
-        ):
-            with self.assertRaisesRegex(
-                ValueError,
-                "POST driver descriptor routes must be implemented as native FastAPI routes",
-            ):
-                native_routes._validate_native_descriptor_driver_routes()
 
     def test_product_onboarding_endpoint_writes_full_launchplane_owned_bundle(self) -> None:
         with TemporaryDirectory() as temporary_directory_name:
@@ -10240,7 +7754,7 @@ class LaunchplaneServiceTests(unittest.TestCase):
             }
         )
         preview_request, generation_request = (
-            control_plane_service._generic_web_preview_refresh_mutation_requests(
+            generic_web_preview_dispatch._generic_web_preview_refresh_mutation_requests(
                 request=GenericWebPreviewRefreshRequest.model_validate(
                     {
                         "schema_version": 1,
